check for file before deletion

+++ functions/update_apps.sh
@@ -43,7 +43,7 @@ do
     fi
 done
 rm temp.txt
-rm trigger
+[[ -e trigger ]] && rm trigger
 echo
 echo
 }
